TrabajoFinal/reservations-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic-settings
psycopg2-binary
pika
python-dotenv
PyJWT
requests
python-json-logger